fix: move CredentialStorage
