+++ lib/app/utils/taskchampion/taskchampion.dart
@@ -140,18 +140,3 @@ class ManageTaskChampionCreds extends StatelessWidget {
     );
   }
 }
-
-class CredentialsStorage {
-  static const String _encryptionSecretKey = 'encryptionSecret';
-  static const String _clientIdKey = 'clientId';
-
-  static Future<String?> getEncryptionSecret() async {
-    SharedPreferences prefs = await SharedPreferences.getInstance();
-    return prefs.getString(_encryptionSecretKey);
-  }
-
-  static Future<String?> getClientId() async {
-    SharedPreferences prefs = await SharedPreferences.getInstance();
-    return prefs.getString(_clientIdKey);
-  }
-}
